:sparkles: [feat] StoreList 조건삭제 추가
