+++ src/components/common/Map/ResponsiveSelectBox.tsx
@@ -117,14 +117,6 @@ export default function ResponsiveSelectBox({
   onRightMoneyMaxChange,
   onStoreCategoryChange,
   onSubStoreCategoryChange,
-  setMainTainDepositMin,
-  setMaintainDepositMax,
-  setMaintainMonthlyMin,
-  setMaintainMonthlyMax,
-  setMaintainRightMoneyMin,
-  setMaintainRightMoneyMax,
-  setMaintainCategory,
-  setMaintainSubCategoryValue,
 }: ResponsiveSelectBoxProps) {
   const [selectedPropertyTypes, setSelectedPropertyTypes] = useState<string[]>(
     []
@@ -167,59 +159,52 @@ export default function ResponsiveSelectBox({
     const value = e.currentTarget.value;
     setDepositMin(value);
     onDepositMinChange(value);
-    setMainTainDepositMin(value);
   };
 
   const handleDepositMaxChange = (e: ChangeEvent<HTMLInputElement>) => {
     const value = e.currentTarget.value;
     setDepositMax(value);
     onDepositMaxChange(value);
-    setMaintainDepositMax(value);
   };
 
   const handleMonthlyMinChange = (e: ChangeEvent<HTMLInputElement>) => {
     const value = e.currentTarget.value;
     setMonthlyMin(value);
     onMonthlyMinChange(value);
-    setMaintainMonthlyMin(value);
   };
 
   const handleMonthlyMaxChange = (e: ChangeEvent<HTMLInputElement>) => {
     const value = e.currentTarget.value;
     setMonthlyMax(value);
     onMonthlyMaxChange(value);
-    setMaintainMonthlyMax(value);
   };
 
   const handleRightMoneyMinChange = (e: ChangeEvent<HTMLInputElement>) => {
     const value = e.currentTarget.value;
     setRightMoneyMin(value);
     onRightMoneyMinChange(value);
-    setMaintainRightMoneyMin(value);
   };
 
   const handleRightMoneyMaxChange = (e: ChangeEvent<HTMLInputElement>) => {
     const value = e.currentTarget.value;
     setRightMoneyMax(value);
     onRightMoneyMaxChange(value);
-    setMaintainRightMoneyMax(value);
   };
 
   const onHandleCategoryChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     const selectedCategory = e.target.value;
     setCategory(selectedCategory);
     onStoreCategoryChange(selectedCategory);
-    setMaintainCategory(selectedCategory);
+
     if (selectedCategory === "전체") {
       setSubCategoryValue("전체");
       onSubStoreCategoryChange("전체");
-      setMaintainSubCategoryValue("전체");
+
       setSubCategory([]);
     } else {
       setSubCategory(subCategories[selectedCategory] || []);
       setSubCategoryValue("전체");
       onSubStoreCategoryChange("전체");
-      setMaintainSubCategoryValue("전체");
     }
   };
   const onHandleSubCategoryChange = (
@@ -228,7 +213,6 @@ export default function ResponsiveSelectBox({
     const selectedSubCategory = e.target.value;
     setSubCategoryValue(selectedSubCategory);
     onSubStoreCategoryChange(selectedSubCategory);
-    setMaintainSubCategoryValue(selectedSubCategory);
   };
 
   return (
